requirements.txt
gradio==4.44.1
python-dotenv==1.0.1
requests==2.32.3
markdown==3.7
beautifulsoup4==4.12.3
pytest==8.2.0
